+++ functions/Prompts.php
@@ -13,6 +13,7 @@
  * Go back in browser history on Cancel (unless $remove_modfunc_on_cancel = false)
  *
  * @since 11.4 Use 'delete_ok' URL param instead of submit button name
+ * @since 12.1 JS If inside colorBox, close it on Cancel
  *
  * @example if ( DeletePrompt( _( 'Title' ) ) ) DBQuery( "DELETE FROM BOK WHERE ID='" . (int) $_REQUEST['benchmark_id'] . "'" );
  *
@@ -54,13 +55,22 @@ function DeletePrompt( $title, $action = 'Delete', $remove_modfunc_on_cancel = t
 
 		PopTable( 'header', _( 'Confirm' ) . ( mb_strpos( $action, ' ' ) === false ? ' ' . $action : '' ) );
 
+		if ( $remove_modfunc_on_cancel )
+		{
+			// If inside colorBox, close it. Otherwise, go back in browser history.
+			$cancel_onclick_js = 'if ($(this).closest("#colorbox").length) $.colorbox.close(); else self.history.go(-1);';
+		}
+		else
+		{
+			$cancel_onclick_js = 'ajaxLink(' . json_encode( $PHP_tmp_SELF_cancel ) . ');';
+		}
+
 		echo '<br><div class="center">' . button( 'warning', '', '', 'bigger' ) .
 			'<h4>' . sprintf( _( 'Are you sure you want to %s that %s?' ), $action, $title ) . '</h4>
 			<form action="' . $PHP_tmp_SELF . '" method="POST">' .
 				SubmitButton( _( 'OK' ), 'delete_ok', '' ) .
 				'<input type="button" name="delete_cancel" class="button-primary" value="' . AttrEscape( _( 'Cancel' ) ) . '"
-					onclick="' . ( $remove_modfunc_on_cancel ? 'javascript:self.history.go(-1);' :
-						AttrEscape( 'ajaxLink(' . json_encode( $PHP_tmp_SELF_cancel ) . ');' ) ) . '">
+					onclick="' . AttrEscape( $cancel_onclick_js ) . '">
 			</form>
 		</div><br>';
 
@@ -83,6 +93,7 @@ function DeletePrompt( $title, $action = 'Delete', $remove_modfunc_on_cancel = t
  * Go back in browser history on Cancel
  *
  * @since 11.4 Use 'delete_ok' URL params instead of submit button name
+ * @since 12.1 JS If inside colorBox, close it on Cancel
  *
  * @example if ( Prompt( _( 'Confirm' ), _( 'Do you want to dance?' ), $message ) )
  *
@@ -109,12 +120,17 @@ function Prompt( $title = 'Confirm', $question = '', $message = '' )
 
 		PopTable( 'header', $title );
 
+		// If inside colorBox, close it. Otherwise, go back in browser history.
+		$cancel_onclick_js = 'if ($(this).closest("#colorbox").length) $.colorbox.close(); else self.history.go(-1);';
+
 		echo '<h4 class="center">' . $question . '</h4>
 			<form action="' . $PHP_tmp_SELF . '" method="POST">' .
 				$message .
 				'<div class="center"><br>' .
 				SubmitButton( _( 'OK' ), 'delete_ok', '' ) .
-				'<input type="button" name="delete_cancel" class="button-primary" value="' . AttrEscape( _( 'Cancel' ) ) . '" onclick="javascript:self.history.go(-1);">
+				// If inside colorBox, close it. Otherwise, go back in browser history.
+				'<input type="button" name="delete_cancel" class="button-primary" value="' . AttrEscape( _( 'Cancel' ) ) .
+					'" onclick="' . AttrEscape( $cancel_onclick_js ) . '">
 				</div>
 			</form><br>';